fix a bug which fails to recognize penalty 0.0 as zero

+++ wrims_v2/wrims_v2/src/wrimsv2/wreslplus/elements/Procedures.java
@@ -334,7 +334,7 @@ public class Procedures {
 
 		try {
 
-			int p = Integer.parseInt(v);
+			float p = Float.parseFloat(v);
 			if (p == 0) return true;
 
 		}
